requirements.txt
azure-ai-inference==1.0.0b9
azure-core==1.32.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
